+++ app/(blog)/blog/blogPage.tsx
@@ -1,43 +1,20 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
 import { PostList } from "@/types/post";
 import { Divider } from "@/components";
-
 import PostBox from "./components/PostBox";
+import TagList from "./components/TagList";
 import * as S from "./style";
 
 export default function BlogContent({ blogList }: { blogList: PostList }) {
-  const router = useRouter();
-  const [tag, setTag] = useState(blogList.searchTag);
   const [isLoading, setIsLoading] = useState(true);
-  const [allTags, setAllTags] = useState<string[]>([]);
 
   useEffect(() => {
-    const fetchTags = async () => {
-      try {
-        const response = await fetch("/api/posts?type=tags");
-        const tags = await response.json();
-        setAllTags(tags);
-      } catch (error) {
-        console.error("태그 로딩 실패:", error);
-      }
-    };
-    fetchTags();
-  }, []);
-
-  useEffect(() => {
-    setTag(blogList.searchTag);
     setTimeout(() => {
       setIsLoading(false);
     }, 500);
   }, [blogList]);
-
-  const handleMove = (newTag: string) => {
-    router.push(`/blog?tag=${newTag}`);
-    setIsLoading(true);
-  };
 
   return (
     <S.ContentBox>
@@ -46,21 +23,12 @@ export default function BlogContent({ blogList }: { blogList: PostList }) {
         <p className="text-gray-600 text-base md:text-lg">공부한 개념을 온전히 이해하기 위해 기록합니다!</p>
       </section>
 
-      <S.MenuTapBox>
-        <S.TapButton isTap={tag === "all"} onClick={() => handleMove("all")}>
-          All
-        </S.TapButton>
-        {allTags.map((uniqueTag) => (
-          <S.TapButton key={uniqueTag} isTap={tag === uniqueTag} onClick={() => handleMove(uniqueTag)} className="mb-2">
-            {uniqueTag}
-          </S.TapButton>
-        ))}
-      </S.MenuTapBox>
+      <TagList currentTag={blogList.searchTag} />
 
       <Divider margin="0 0 16px 0" />
 
       <S.TapTitle>
-        📚 {tag === "all" ? "All" : tag} ({blogList?.totalElement})
+        📚 {blogList.searchTag === "all" ? "All" : blogList.searchTag} ({blogList?.totalElement})
       </S.TapTitle>
 
       <S.PostContainer>
@@ -69,7 +37,7 @@ export default function BlogContent({ blogList }: { blogList: PostList }) {
         ) : (blogList?.data.length || 0) > 0 ? (
           blogList?.data.map((post) => <PostBox key={post._id} post={post} />)
         ) : (
-          <div className="text-center py-12 text-gray-500">{tag}에 해당하는 포스트가 없어요</div>
+          <div className="text-center py-12 text-gray-500">{blogList.searchTag}에 해당하는 포스트가 없어요</div>
         )}
       </S.PostContainer>
     </S.ContentBox>
